# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LOGIK GAME API",
        "description": "API backend du jeu télévisé LOGIK GAME — gestion des sessions, manches, questions, joueurs, temps réel et projection.",
        "contact": {
            "name": "NerdX",
            "email": "contact@nerdx.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/admin/sessions/{session}/rounds/{round}/questions/{question}/hint": {
            "get": {
                "tags": [
                    "Question Hints"
                ],
                "summary": "Récupérer l'indice d'une question",
                "operationId": "52f6fadf717b1354dab8697f20eb1d4c",
                "parameters": [
                    {
                        "name": "session",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "round",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "question",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Indice de la question"
                    },
                    "404": {
                        "description": "Aucun indice configuré"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Question Hints"
                ],
                "summary": "Créer ou remplacer l'indice d'une question",
                "description": "Uniquement pour les questions des manches de type \"hint\" (manche 2). Crée l'indice s'il n'existe pas, le remplace sinon.",
                "operationId": "91541ff387146363d0d494e05e6ea7be",
                "parameters": [
                    {
                        "name": "session",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "round",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "question",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "hint_type"
                                ],
                                "properties": {
                                    "hint_type": {
                                        "description": "Type d'indice selon le type de question",
                                        "type": "string",
                                        "enum": [
                                            "remove_choices",
                                            "reveal_letters",
                                            "reduce_range"
                                        ]
                                    },
                                    "time_penalty_seconds": {
                                        "description": "Secondes retirées du timer quand l'indice est activé",
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "removed_choice_ids": {
                                        "description": "IDs des choix à masquer (hint_type=remove_choices)",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "nullable": true
                                    },
                                    "revealed_letters": {
                                        "description": "Positions des lettres révélées (hint_type=reveal_letters)",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "nullable": true
                                    },
                                    "range_hint_text": {
                                        "description": "Texte d'indice libre (hint_type=range), ex: \"Entre 50 et 100\"",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "range_min": {
                                        "type": "number",
                                        "nullable": true
                                    },
                                    "range_max": {
                                        "type": "number",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Indice mis à jour"
                    },
                    "422": {
                        "description": "Validation échouée ou manche incompatible"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Question Hints"
                ],
                "summary": "Supprimer l'indice d'une question",
                "operationId": "642b93d7b23e927ab7616ed765de3715",
                "parameters": [
                    {
                        "name": "session",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "round",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "question",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Indice supprimé"
                    },
                    "404": {
                        "description": "Aucun indice à supprimer"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token Sanctum admin (obtenu via POST /admin/login)",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            },
            "playerToken": {
                "type": "apiKey",
                "description": "Token d'accès joueur (64 caractères, reçu après sélection)",
                "name": "X-Player-Token",
                "in": "header"
            }
        }
    },
...
}